backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for Asviglager asset management system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "motor>=3.6.0",
    "pymongo>=4.10.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.0",
]
